Refactored Node build to install into a “node” directory
