+++ .gitignore
@@ -9,5 +9,6 @@ install_manifest.txt
 CTestTestfile.cmake
 .DS_Store
 build*/
+node/
 node_modules/
 pod_deploy/
