+++ node_install.sh
@@ -1,4 +1,5 @@
-mkdir build_node
-cd build_node
-cmake -DBUILD_NODE=ON -DCMAKE_POSITION_INDEPENDENT_CODE=true -DCMAKE_BUILD_TYPE=Release ../
+mkdir -p node/build
+cd node/build
+cmake -DBUILD_NODE=ON -DCMAKE_POSITION_INDEPENDENT_CODE=true -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=../ ../../ 
 make
+make install
