+++ src/wickrcrypto/swig/CMakeLists.txt
@@ -20,3 +20,7 @@ endif()
 if(BUILD_JAVA)
     add_subdirectory(java)
 endif()
+
+# Install the SWIG files so dependencies can utilize them
+file(GLOB SwigFiles *.i)
+install(FILES ${SwigFiles} DESTINATION share/wickrcrypto/swig)
