+++ src/wickrcrypto/swig/node/test/deviceInfoTests.js
@@ -2,7 +2,7 @@
 'use strict'
 
 var expect = require('expect.js')
-var wickrcrypto = require('../../../../../build_node/src/wickrcrypto/swig/node/lib/wickrcrypto')
+var wickrcrypto = require('../../../../../node/lib/wickrcrypto')
 
 describe ("Device Info Tests", function() {
     it ("should be able to create device info", function() {
