+++ src/wickrcrypto/swig/node/test/nodeTests.js
@@ -1,7 +1,7 @@
 'use strict'
 
 var expect = require('expect.js')
-var wickrcrypto = require('../../../../../build_node/src/wickrcrypto/swig/node/lib/wickrcrypto')
+var wickrcrypto = require('../../../../../node/lib/wickrcrypto')
 
 describe ("Node Tests", function() {
     it ("should be able to create a node", function() {
