+++ src/wickrcrypto/swig/node/test/storageKeyTests.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var expect = require('expect.js')
-var wickrcrypto = require('../../../../../build_node/src/wickrcrypto/swig/node/lib/wickrcrypto')
+var wickrcrypto = require('../../../../../node/lib/wickrcrypto')
 
 describe ("Storage Keys Tests", function() {
     it ("should be able to create storage keys", function() {
